move mock objects creation to a new package
